docker: moved fingerprint code to it's own file
